Rename DOM utilities file to make more sense.
